requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
neo4j==5.14.0
psycopg2-binary==2.9.9
chromadb==0.4.24
httpx==0.25.2
networkx==3.2.1
python-multipart==0.0.6
